+++ model/src/main/java/model/ShapefileWKT.java
@@ -22,9 +22,9 @@ public class ShapefileWKT {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private long id;
 
-	@Column(length=10000)
+	@Column(length = 10000)
 	private String wkt;
 
 	@Column(unique = true)
@@ -43,16 +43,15 @@ public class ShapefileWKT {
 	}
 
 	public void setShapefileType(String shapefile) {
-		if (shapefile == CORNER
-				|| shapefile == CORNER || shapefile == BUS_MAXIMAL
-				|| shapefile == BUS_STOP || shapefile == ADDRESS
-				|| shapefile == STREET || shapefile == CORNER 
-				|| shapefile == OBSTACLE
+		if (shapefile == CORNER || shapefile == CORNER
+				|| shapefile == BUS_MAXIMAL || shapefile == BUS_STOP
+				|| shapefile == ADDRESS || shapefile == STREET
+				|| shapefile == CORNER || shapefile == OBSTACLE
 				|| shapefile == CONTROL_POINT)
 			this.shapefileType = shapefile;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 }
